requirements.txt
streamlit
pyrebase4
pandas
numpy
matplotlib
seaborn
setuptools
time
io